updated verifier for better output, removed key generation from player since verifier should be run first
